Add some missing feats and class abilities
Deadly Defense and the Dervish's Fast Movement, plus the Dervish's AC
bonus.  I prefixed thelatter with "Dervish" in the text-to-match,
since (a) it's too generic and (b) that's what I've written on Penny's
character sheet (essentially because (a) applies there, too).  After
Anders pointed out Fast Movement is a class feature of a few other
classes, I did the same for it; and added the one for barbarians.

Checking other character sheets in our campaign, I then extended this
to all the other feats, class features and special abilities I found
there without links.