Rename a variable and correct a misleading comment
The helpTable maps initial fragments of names to URLs; so using URL as
name for the keys was misleading. The code in fact checked for whether
the key was a prefix of the name to be looked up, not the other way
round as claimed by the comment (quite apart from describing the key
as an URL).  Save the next reader the confusion I've just had to
disentangle by describing the key and check correctly.  Replace an out
of date comment with a description of the helpTable.