Enable more rows in the Weapon and Armor tables
A character with diverse different ways to attack can easily have more
than four weapons. Since the table to enable new rows aligns Armor
with Weapons, include as meany for Protective Items as for Weapons,
even though that may be less relevant. Allow the possibility of a
secondary shield as one of the new Armor slots.

This naturally widens the space these enablers take up, so widen their
cells of the table in which they appear, separating new class enablers
from their old class wide. The other two class=wide columns in the
table have ample room (their content is a number unlikely to make it
to three digits, much less more; so their titles are the constraints
on width), so we can pilfer 5% of total width off each of those and
give it to the enablers' column.

In the process, move the styling of the check-boxes from a style
attribute on each of them to a suitable CSS rule; and add one more
&nbsp; to the Armor label to (hopefully) fix the alignment of the
checkboxes in the two rows.  (I, at least, see the Armor row's
check-boxes visible to the left of the Weapon row's.)  Also clear out
some trash (blank lines, stray comment) and make some doc comment a
bit more truthful / coherent.

Since enablers share their column with the spell resistance entry, we
can now give that its full name instead of truncating to SPELL RESIST.