Fix typo in name of weapon-crit cell name
When someone converted the code from dealing with each of Weapon1Crit,
Weapon2Crit and Weapon3Crit independently to using "Weaopon" + i +
"1Crit" they rather obviously failed to delete the 1 from the start of
that last. Do so now, but include a kludge-around so that those with
data saved with the typo do get it looked up with the typo'd name (but
eventually stored with the right name).